// src/api/generated/models/categoriesTreeParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { CategoriesTreeLevel } from './categoriesTreeLevel';
import type { CategoriesTreeSortBy } from './categoriesTreeSortBy';
import type { CategoriesTreeSortOrder } from './categoriesTreeSortOrder';
import type { CategoriesTreeType } from './categoriesTreeType';

export type CategoriesTreeParams = {
search?: string;
type?: CategoriesTreeType;
isActive?: boolean;
level?: CategoriesTreeLevel;
parentId?: string;
sortBy?: CategoriesTreeSortBy;
sortOrder?: CategoriesTreeSortOrder;
};